move colorpad into colorpicker module
