+++ rnote-ui/src/main.rs
@@ -15,7 +15,6 @@ mod appwindow;
 mod canvas;
 mod canvasmenu;
 mod canvaswrapper;
-mod colorpad;
 mod colorpicker;
 mod iconpicker;
 mod mainheader;
@@ -32,7 +31,6 @@ pub(crate) use appwindow::RnoteAppWindow;
 pub(crate) use canvas::RnoteCanvas;
 pub(crate) use canvasmenu::CanvasMenu;
 pub(crate) use canvaswrapper::RnoteCanvasWrapper;
-pub(crate) use colorpad::ColorPad;
 pub(crate) use colorpicker::ColorPicker;
 pub(crate) use iconpicker::IconPicker;
 pub(crate) use mainheader::MainHeader;
