+++ rnote-ui/src/meson.build
@@ -61,6 +61,7 @@ rnote_ui_sources = files(
     'canvas/input.rs',
     'canvas/imexport.rs',
     'colorpicker/colorsetter.rs',
+    'colorpicker/colorpad.rs',
     'colorpicker/mod.rs',
     'penssidebar/brushpage.rs',
     'penssidebar/eraserpage.rs',
@@ -97,7 +98,6 @@ rnote_ui_sources = files(
     'config.rs',
     'globals.rs',
     'iconpicker.rs',
-    'colorpad.rs',
     'main.rs',
     'mainheader.rs',
     'unitentry.rs',
